requirements.txt
coverage==5.3
grpcio==1.34.0
grpcio-testing==1.34.0
grpcio-tools==1.33.2
protobuf==3.14.0
pycodestyle==2.6.0
six==1.15.0
PyYAML==5.3.1